post

+++ app/Http/Controllers/TasksController.php
@@ -6,6 +6,9 @@ use Illuminate\Http\Request;
 
 use App\Task;
 
+use App\User;
+
+
 class TasksController extends Controller
 {
     /**
@@ -16,10 +19,19 @@ class TasksController extends Controller
     public function index()
     {
         $tasks = Task::all();
-
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks= $user->tasks()->orderBy('created_at','desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        return view('welcome',$data);
+        
     }
 
     /**
@@ -44,15 +56,16 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
+         $this->validate($request, [
             'status' => 'required|max:10', 
         ]);
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
-        return redirect('/');
+        
+        
+        $request->user()->tasks()->create([
+            'content'=>$request->content,
+            'status'=>$request->status,
+            ]);
+            return redirect('/');
     }
 
     /**
@@ -63,11 +76,21 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
+        $task =Task::find($id);
+        return view('tasks.show',[
+            'task'=>$task,
+            ]);
+        // $user = User::find($id);
+        // $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        // $data = [
+        //     'user' => $user,
+        //     'tasks' => $tasks,
+        // ];
+
+        // $data += $this->counts($user);
+
+        // return view('tasks.show', $data);
     }
 
     /**
@@ -113,9 +136,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
+
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         return redirect('/');
     }
+    
 }
